Since server can now change FPS on the fly, client can't assume Setup->frames_per_second (FPS) is the current value of FPS.
The client now uses clientFPS (the number of frames the client drew last
second) for estimating the correct number of frames to do the base warning
(the correct time to use for base warning is assumed to be 3 seconds).

Also changed max. frequency of blinking for stuff like base warning,
score objects etc. to be 20 Hz.