Added code into name matching function Get_player_by_name() so that if the given string matches exactly the beginning of only one nick, then that player is returned, even if the given string would be a substring of some other nick. The idea is that if we have players Alpha and Beta then if we write a message
a: hi!

it will go to Alpha. The substring matching part that comes last
in the function Get_player_by_name() is mostly useful if two players
nicks have a common first part, we can match on the second part.
E.g. if we have players OpticalMouse and OpticalCop we can do

mo: hello

and the message will go to OpticalMouse. With the 4.5.4 matching
code you would have had to write

opticalm: hello

Same goes for server commands.