Confmenu now starts with online color indicators, click the colored square and it changes into a color setup widget (roughly 4 lines high)
Now the confmenu is persistant from start 'til finish since there is some
serious bug with closing it =( couldn't find what was going on.
Good side is that the cnf menu now remembers its position (i.e. scroll
position) and wich colors were edited this session (their widgets are
still expanded)
Bad side is that the entire conf menu is persistant in memory, and uses
up alot of texture space =(
Atleast *I* can't make it crash now, but the trick to crash it seems diff
erent between machines (col's method didn't crash it on my comp, but on
my comp it started crapping stuff up if a message appeared while confmenu
was up, and I closed with the button)

/virus